5 years stock data
Using this code, we can fetch the last 5 years of data for a particular stock. However, the drawback is that we need to explicitly write the stock symbol to retrieve data for different stocks. The code provides data for the last 5 years, but you can change the timeline to suit your requirements.